accelerators/sharepoint-connector/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --extra func --no-dev
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.38.2
    # via
    #   azure-identity
    #   azure-search-documents
azure-functions==1.24.0
    # via sharepoint-connector
azure-identity==1.25.2
    # via sharepoint-connector
azure-search-documents==11.6.0
    # via sharepoint-connector
beautifulsoup4==4.13.5
    # via
    #   extract-msg
    #   sharepoint-connector
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via tqdm
colorclass==2.2.2
    # via oletools
compressed-rtf==1.0.7
    # via extract-msg
cryptography==46.0.5
    # via
    #   azure-identity
    #   msal
    #   msoffcrypto-tool
    #   pyjwt
defusedxml==0.7.1
    # via odfpy
distro==1.9.0
    # via openai
easygui==0.98.3
    # via oletools
ebcdic==1.1.1
    # via extract-msg
ebooklib==0.20
    # via sharepoint-connector
et-xmlfile==2.0.0
    # via openpyxl
extract-msg==0.55.0
    # via sharepoint-connector
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   sharepoint-connector
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
isodate==0.7.2
    # via azure-search-documents
jiter==0.13.0
    # via openai
lark==1.3.1
    # via rtfde
lxml==6.0.2
    # via
    #   ebooklib
    #   python-docx
    #   python-pptx
markupsafe==3.0.3
    # via werkzeug
msal==1.35.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msoffcrypto-tool==6.0.0 ; (platform_python_implementation != 'PyPy' and sys_platform == 'darwin') or (platform_python_implementation != 'PyPy' and sys_platform == 'win32') or (sys_platform != 'darwin' and sys_platform != 'win32')
    # via oletools
odfpy==1.4.1
    # via sharepoint-connector
olefile==0.47
    # via
    #   extract-msg
    #   msoffcrypto-tool
    #   oletools
oletools==0.60.2
    # via
    #   pcodedmp
    #   rtfde
openai==2.24.0
    # via sharepoint-connector
openpyxl==3.1.5
    # via sharepoint-connector
pcodedmp==1.2.6
    # via oletools
pillow==12.1.1
    # via python-pptx
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.11.0
    # via msal
pymupdf==1.27.1
    # via sharepoint-connector
pyparsing==3.3.2
    # via oletools
python-docx==1.2.0
    # via sharepoint-connector
python-dotenv==1.2.1
    # via sharepoint-connector
python-pptx==1.0.2
    # via sharepoint-connector
red-black-tree-mod==1.22
    # via extract-msg
requests==2.32.5
    # via
    #   azure-core
    #   msal
rtfde==0.1.2.2
    # via extract-msg
six==1.17.0
    # via ebooklib
sniffio==1.3.1
    # via openai
soupsieve==2.8.3
    # via beautifulsoup4
striprtf==0.0.29
    # via sharepoint-connector
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   azure-core
    #   azure-identity
    #   azure-search-documents
    #   beautifulsoup4
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3 ; sys_platform == 'win32'
    # via tzlocal
tzlocal==5.3.1
    # via extract-msg
urllib3==2.6.3
    # via requests
werkzeug==3.1.6
    # via azure-functions
win-unicode-console==0.5 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
    # via pcodedmp
xlsxwriter==3.2.9
    # via python-pptx
